api_daily_tasks/requirements.txt
django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
markdown==3.3.6
python-decouple==3.6
unipath==1.1
